Pull out TransportTests base suite
